fix: index only entities related to paginable record BOs
