+++ src/main/java/biblivre/cataloging/HttpRequestSearchHelper.java
@@ -1,21 +1,22 @@
 package biblivre.cataloging;
 
 import biblivre.administration.indexing.IndexingGroups;
+import biblivre.cataloging.bibliographic.PaginableRecordBO;
 import biblivre.cataloging.search.SearchDTO;
 import biblivre.core.ExtendedRequest;
 
 public class HttpRequestSearchHelper {
-    public static SearchDTO paginate(ExtendedRequest request, RecordBO recordBO) {
+    public static SearchDTO paginate(ExtendedRequest request, PaginableRecordBO paginableRecordBO) {
         Integer searchId = request.getInteger("search_id", null);
 
-        SearchDTO search = recordBO.getSearch(searchId);
+        SearchDTO search = paginableRecordBO.getSearch(searchId);
 
         if (search == null) {
             return null;
         }
 
         Integer defaultSortableGroupId =
-                IndexingGroups.getDefaultSortableGroupId(recordBO.getRecordType());
+                IndexingGroups.getDefaultSortableGroupId(paginableRecordBO.getRecordType());
 
         Integer sort = request.getInteger("sort", defaultSortableGroupId);
